Validações customizadas + subtotal + total do pedido

+++ src/main/java/br/com/sale/domain/Category.java
@@ -19,7 +19,7 @@ public class Category implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 	
 
@@ -29,17 +29,17 @@ public class Category implements Serializable {
 	public Category() {
 	}
 
-	public Category(Long id, String nome) {
+	public Category(Integer id, String nome) {
 		super();
 		this.id = id;
 		this.nome = nome;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
